Feature/gitgud search endpoint

+++ src/main/java/com/gitgud/api/objects/ApiSearchResults.java
@@ -4,25 +4,25 @@ import java.util.List;
 
 public class ApiSearchResults {
 
-    private int maxPrice;
-    private int minPrice;
+    private long maxPrice;
+    private long minPrice;
     private double maxSize;
     private double minSize;
     private List<ApiRealState> elements;
 
-    public int getMaxPrice() {
+    public long getMaxPrice() {
         return maxPrice;
     }
 
-    public void setMaxPrice(int maxPrice) {
+    public void setMaxPrice(long maxPrice) {
         this.maxPrice = maxPrice;
     }
 
-    public int getMinPrice() {
+    public long getMinPrice() {
         return minPrice;
     }
 
-    public void setMinPrice(int minPrice) {
+    public void setMinPrice(long minPrice) {
         this.minPrice = minPrice;
     }
